URL Handler support as per OSGi Specification.
This makes all the org.osgi.test.cases.url TCK tests pass.

The code works both in AS7 as well as in standalone mode, but for
standalone a number of hacks were put in place to get the ServiceLoader
mechanism to work. The changes to:
  FrameworkLocalLoader
  SystemLocalLoader
  ModuleManagerPluginImpl
  SystemPackagesPluginImpl
and STANDALONE block in URLHandlerPluginImpl are all put in purely to
get this to work in standalone and are not needed when running in AS7.
The problem addressed by these changes is isolated in the
FrameworkServiceLoaderTestCase.

Also included additional new unit tests. Still to do: system tests.